chore: remove ENABLE_NEW_ARTWORK_FILTERS feature flag (#7771)
* Showed new filters, merged aggregations and additionalAggregations arrays, SizeFilter2 renamed to SizeFilter, deleted old SizeFilter

* Merged tests

* Removed ENABLE_NEW_ARTWORK_FILTERS variable

* Removed @ts-expect-error STRICT_NULL_CHECK comments from SizeFilter.tsx

* Deleted empty Text component
