+++ src/lib/setup_sharify.js
@@ -48,7 +48,6 @@ sharify.data = _.extend(
     "ENABLE_REQUEST_CONDITION_REPORT",
     "ENABLE_SERVER_SIDE_CACHE",
     "ENABLE_SIGN_IN_WITH_APPLE",
-    "ENABLE_NEW_ARTWORK_FILTERS",
     "ENABLE_QUERY_BATCHING",
     "ENABLE_WEB_CRAWLING",
     "EOY_2016_ARTICLE",
