+++ src/typings/sharify.d.ts
@@ -26,7 +26,6 @@ declare module "sharify" {
       CURRENT_USER: User
       readonly DEPLOY_ENV: string
       readonly EIGEN: boolean
-      readonly ENABLE_NEW_ARTWORK_FILTERS: boolean
       readonly ENABLE_SERVER_SIDE_CACHE: string
       readonly ENABLE_SIGN_IN_WITH_APPLE: string // TODO: Remove once sign in with apple is rolled out.
       readonly ENABLE_QUERY_BATCHING: boolean
